+++ app/src/androidTest/java/com/example/a3rd_seminar/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.sopt_study
+package com.example.a3rd_seminar
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.sopt_study", appContext.packageName)
+        assertEquals("com.example.a3rd_seminar", appContext.packageName)
     }
 }
