+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.sopt_study">
+    package="com.example.a3rd_seminar">
 
     <application
         android:allowBackup="true"
@@ -8,19 +8,12 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Sopt_Study">
-        <activity
-            android:name=".HomeActivity"
-            android:exported="false"
-            android:label="@string/title_activity_home"
-            android:theme="@style/Theme.Sopt_Study.NoActionBar" />
+        android:theme="@style/Theme.3rd_seminar">
         <activity
             android:name=".SignUpActivity"
-            android:exported="false"
-            android:label="@string/title_activity_sign_up"
-            android:theme="@style/Theme.Sopt_Study.NoActionBar" />
+            android:exported="false" />
         <activity
-            android:name=".EmptyActivity"
+            android:name=".HomeActivity"
             android:exported="false" />
         <activity
             android:name=".SignInActivity"
