+++ app/src/main/java/com/example/a3rd_seminar/HomeActivity.kt
@@ -1,0 +1,16 @@
+package com.example.a3rd_seminar
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.example.a3rd_seminar.databinding.ActivityHomeBinding
+
+class HomeActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_sign_up)
+
+
+
+    }
+}
