+++ app/src/main/java/com/example/a3rd_seminar/SignInActivity.kt
@@ -1,0 +1,14 @@
+package com.example.a3rd_seminar
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.example.a3rd_seminar.databinding.ActivitySignInBinding
+
+class SignInActivity : AppCompatActivity() {
+    private lateinit var binding : ActivitySignInBinding
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivitySignInBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+    }
+}
