+++ app/src/main/java/com/example/a3rd_seminar/SignUpActivity.kt
@@ -1,0 +1,11 @@
+package com.example.a3rd_seminar
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+
+class SignUpActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_sign_up)
+    }
+}
