--- app/src/main/java/com/example/sopt_study/FollowerAdapter.kt
@@ -1,36 +1,0 @@
-package com.example.sopt_study
-
-import android.view.LayoutInflater
-import androidx.recyclerview.widget.RecyclerView
-import android.view.ViewGroup
-import com.example.sopt_study.databinding.UserListBinding
-
-
-class FollowerAdapter : RecyclerView.Adapter<FollowerAdapter.FollowerViewHolder>() {
-
-    val userList = mutableListOf<UserData>()
-    //질문하기 :: private
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): FollowerViewHolder {
-        val binding =
-            UserListBinding.inflate(LayoutInflater.from(parent.context), parent, false)
-        return FollowerViewHolder(binding)
-    }
-
-    override fun onBindViewHolder(holder: FollowerViewHolder, position: Int) {
-        holder.onBind(userList[position])
-    }
-
-    override fun getItemCount(): Int = userList.size
-
-    class FollowerViewHolder(
-        private val binding: UserListBinding
-    ) : RecyclerView.ViewHolder(binding.root) {
-        fun onBind(data: UserData) {
-            binding.textName.text = data.name
-            binding.textIntroduce.text = data.introduction
-        }
-    }
-
-
-}
