--- app/src/main/java/com/example/sopt_study/UserData.kt
@@ -1,6 +1,0 @@
-package com.example.sopt_study
-
-data class UserData(
-    val name : String,
-    val introduction : String
-)
