+++ app/src/main/res/drawable/default_edit_text_background.xml
@@ -1,0 +1,47 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item android:state_focused="true">
+
+        <shape android:shape="rectangle">
+
+            <stroke
+
+                android:width="1dp"
+
+                android:color="#8040FF" />
+
+            <corners android:radius="4dp" />
+
+            <solid
+
+                android:color="#FFFFFF" />
+
+        </shape>
+
+
+    </item>
+
+    <item android:state_focused="false">
+
+        <shape android:shape="rectangle">
+
+            <stroke
+
+                android:width="1dp"
+
+                android:color="#FFFFFF" />
+
+            <corners android:radius="4dp" />
+
+            <solid
+
+                android:color="#F6F5F9" />
+
+        </shape>
+
+
+    </item>
+
+</selector>
