--- app/src/main/res/drawable/round_angle.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-    <item>
-
-        <shape android:shape="rectangle">
-            <stroke android:width="2dp"
-                android:color="@color/black"/>
-            <solid android:color="#ffffff"/>
-            <corners
-                android:bottomLeftRadius="12dp"
-                android:topRightRadius="12dp"
-                android:bottomRightRadius="12dp"
-                android:topLeftRadius="12dp"/>
-        </shape>
-
-    </item>
-</layer-list>
