+++ app/src/main/res/drawable/round_angle_button.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <shape android:shape="rectangle">
+
+
+        <corners android:radius="4dp" />
+
+        <solid
+
+            android:color="#8040FF" />
+
+    </shape>
+
+
+</selector>
