--- app/src/main/res/layout/activity_main.xml
@@ -1,98 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".SignInActivity">
-
-    <Button
-        android:id="@+id/logInButton"
-        android:layout_width="346dp"
-        android:layout_height="58dp"
-        android:layout_marginTop="30dp"
-        android:text="LOGIN"
-
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/editText_password" />
-
-    <Button
-        android:id="@+id/signUpButton"
-        android:layout_width="346dp"
-        android:layout_height="58dp"
-        android:layout_marginTop="30dp"
-        android:backgroundTint="@color/my_green"
-        android:text="회원가입"
-
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/logInButton"
-        app:layout_constraintVertical_bias="0.0" />
-
-    <TextView
-        android:id="@+id/textView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="80dp"
-        android:text="SOPT"
-        android:textSize="60sp"
-
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/textView_ID"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="30dp"
-        android:layout_marginTop="40dp"
-        android:text="아이디"
-        android:textSize="20dp"
-        android:textStyle="bold"
-
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView" />
-
-    <EditText
-        android:id="@+id/editText_id"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginHorizontal="25dp"
-        android:layout_marginTop="10dp"
-
-        android:hint="아이디를 입력해주세요."
-        app:layout_constraintTop_toBottomOf="@+id/textView_ID"
-        tools:layout_editor_absoluteX="20dp" />
-
-    <TextView
-        android:id="@+id/textView_PassWord"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="30dp"
-        android:layout_marginTop="40dp"
-        android:text="비밀번호"
-        android:textSize="20dp"
-        android:textStyle="bold"
-
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/editText_id" />
-
-    <EditText
-        android:id="@+id/editText_password"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginHorizontal="25dp"
-        android:layout_marginTop="10dp"
-        android:hint="비밀번호를 입력해 주세요."
-        android:inputType="textPassword"
-
-        app:layout_constraintTop_toBottomOf="@+id/textView_PassWord"
-        tools:layout_editor_absoluteX="20dp" />
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
