--- app/src/main/res/layout/fragment_follower_page.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".FollowerPageFragment">
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rv_follower"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-
-        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-        tools:itemCount="4"
-        tools:listitem="@layout/user_list" />
-</FrameLayout>
