--- app/src/main/res/layout/fragment_repository_page.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".RepositoryPageFragment">
-
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rv_repository"
-
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-
-        app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
-        app:spanCount="2"
-
-        tools:itemCount="4"
-        tools:listitem="@layout/repository_list" />
-
-</FrameLayout>
