--- app/src/main/res/values-w1240dp/dimens.xml
@@ -1,3 +1,0 @@
-<resources>
-    <dimen name="fab_margin">200dp</dimen>
-</resources>
