--- app/src/main/res/values/dimens.xml
@@ -1,3 +1,0 @@
-<resources>
-    <dimen name="fab_margin">16dp</dimen>
-</resources>
