--- app/src/main/res/values/myColor.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-
-    <color name = "my_green">#78C257</color>
-
-</resources>
