+++ app/src/main/res/values/some_id.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string
+        name="some_id"
+
+        >깃허브 아이디</string>
+</resources>
