+++ app/src/main/res/values/strings.xml
@@ -1,15 +1,6 @@
 <resources>
-    <string name="app_name">Sopt_Study</string>
-    <string name="title_activity_sign_up">SignUpActivity</string>
-    <!-- Strings used for fragments for navigation -->
-    <string name="first_fragment_label">First Fragment</string>
-    <string name="second_fragment_label">Second Fragment</string>
-    <string name="next">Next</string>
-    <string name="previous">Previous</string>
-
-    <string name="hello_first_fragment">Hello first fragment</string>
-    <string name="hello_second_fragment">Hello second fragment. Arg: %1$s</string>
-    <string name="title_activity_home">HomeActivity</string>
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="app_name">3rd_seminar</string>
+    <string name="string_password">비밀번호</string>
+    <string name="login_button_string">로그인</string>
+    <string name="password_button_string">비밀번호</string>
 </resources>
