+++ settings.gradle
@@ -12,5 +12,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "Sopt_Study"
+rootProject.name = "3rd_seminar"
 include ':app'
